fix(sync permissions): extract to generic function
